+++ src/TRestDetectorSingleChannelAnalysisProcess.cxx
@@ -48,9 +48,9 @@ void TRestDetectorSingleChannelAnalysisProcess::InitProcess() {
     } else {
         for (int i = 0; i < fReadout->GetNumberOfReadoutPlanes(); i++) {
             auto plane = fReadout->GetReadoutPlane(i);
-            for (int j = 0; j < plane->GetNumberOfModules(); j++) {
+            for (size_t j = 0; j < plane->GetNumberOfModules(); j++) {
                 auto mod = plane->GetModule(j);
-                for (int k = 0; k < mod->GetNumberOfChannels(); k++) {
+                for (size_t k = 0; k < mod->GetNumberOfChannels(); k++) {
                     auto channel = mod->GetChannel(k);
                     fChannelGain[channel->GetDaqID()] = 1;       // default correction factor is 1
                     fChannelGainError[channel->GetDaqID()] = 1;  // relative error
